Proposal: cTLS without reconstruction
This is a radical change that attempts to implement the (possible)
consensus to move from a "reconstruction-based" transcript to a
"non-reconstruction" transcript.  Verifying the transcript without
reconstruction requires the transcript to include the template,
which requires a byte-identical representation of the template.
This change makes this possible by introducing a stable binary
representation of the template, with JSON as a human-readable
alternative.

This change also attempts to:
* Clarify the use of cTLS in datagram contexts
 - Including how retransmission works without sequence numbers
* Formalize the notion of when vector lengths can be omitted
* Rearrange the extension templating in a more convenient fashion

Fixes #57